move FileType from utils.py to types.py
